# target/classes/openAPI/openAPI.yaml
openapi: 3.0.3
info:
  title: Hotel Management API
  description: API for managing hotels, including searching, retrieving, and creating hotel records.
  version: 1.0.0
servers:
  - url: http://localhost/v1/property-view
paths:
  /hotels:
    get:
      tags:
        - HotelController
      summary: Get a list of all hotels
      operationId: getHotels
      parameters:
        - name: page
          in: query
          required: false
          description: The page number to retrieve (default is 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: The number of items to retrieve per page (default is 10)
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: A list of hotels with brief information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelBrief'
        '400':
          description: Bad Request if the parameters are invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    post:
      tags:
        - HotelController
      summary: Create a new hotel
      operationId: createHotel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewHotel'
      responses:
        '201':
          description: Hotel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelBrief'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  /hotels/{id}:
    get:
      tags:
        - HotelController
      summary: Get detailed information about a specific hotel
      operationId: getHotelById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the hotel
          schema:
            type: integer
      responses:
        '200':
          description: Detailed information about a hotel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelDetail'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  /search:
    get:
      tags:
        - SearchController
      summary: Search hotels by various parameters
      operationId: searchHotels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelSearchCriteriaDTO'
      responses:
        '200':
          description: A list of hotels matching the search criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelBrief'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  /hotels/{id}/amenities:
    post:
      tags:
        - AmenitiesController
      summary: Add amenities to a hotel
      operationId: addHotelAmenities
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the hotel
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '204':
          description: Amenities added successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  /histogram/{param}:
    get:
      tags:
        - SearchController
      summary: Get count of hotels grouped by specified parameter
      operationId: getHotelHistogram
      parameters:
        - name: param
          in: path
          required: true
          description: Parameter to group by (brand, city, country, amenities)
          schema:
            type: string
      responses:
        '200':
          description: Count of hotels grouped by parameter
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
                  format: int64
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

components:
  schemas:
    HotelBrief:
      $ref: "hotel-brief-dto.yaml#/components/schemas/HotelBrief"
    HotelDetail:
      $ref: "hotel-details-dto.yaml#/components/schemas/HotelDetail"
    NewHotel:
      $ref: "new-hotel-dto.yaml#/components/schemas/NewHotel"
    HotelSearchCriteriaDTO:
      $ref: "./hotel-search-criteria-dto.yaml#/components/schemas/HotelSearchCriteriaDTO"
    HistogramDTO:
      $ref: "./histogram-dto.yaml#/components/schemas/HistogramDTO"
    ErrorDto:
      $ref: "error-dtos.yaml#/components/schemas/ErrorDto"